fix: whitelist handle + failing tests #1

+++ packages/votemarket/src/interfaces/IHook.sol
@@ -3,5 +3,5 @@ pragma solidity 0.8.19;
 
 interface IHook {
     function validateHook() external view returns (bool);
-    function doSomething(uint campaignId, uint epoch, bytes calldata) external;
+    function doSomething(uint256 campaignId, uint256 epoch, bytes calldata) external;
 }
